requirements.txt
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.0
albumentations>=1.4.0
Pillow>=10.0.0
numpy>=1.24.0

